src/third_party/blink/web_tests/html/document_metadata/referrer-comma-separated-use-counter.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<head>
  <meta name=referrer content="origin, origin" />
</head>
<script>
  'use strict';

  test(() => {
    const kHTMLMetaElementReferrerPolicyOutsideHead = 3314; // from web_feature.mojom
    const kHTMLMetaElementReferrerPolicyMultipleTokens = 3315;
    assert_false(internals.isUseCounted(document, kHTMLMetaElementReferrerPolicyOutsideHead));
    assert_true(internals.isUseCounted(document, kHTMLMetaElementReferrerPolicyMultipleTokens));
  }, '<meta name=referrer> with multiple tokens is use counted');
</script>
